Bmi correction

+++ lib/src/features/auth/presentation/controllers/auth_controller.dart
@@ -1,6 +1,14 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:nishauri/src/features/appointments/data/repository/appointment_repository.dart';
+import 'package:nishauri/src/features/appointments/data/services/appointment_services.dart';
 import 'package:nishauri/src/features/auth/data/models/auth_state.dart';
 import 'package:nishauri/src/features/auth/data/respositories/auth_repository.dart';
+import 'package:nishauri/src/features/clinic_card/data/repository/ProgramRepository.dart';
+import 'package:nishauri/src/features/clinic_card/data/services/ProgramService.dart';
+import 'package:nishauri/src/features/hiv/data/repositories/art_appointment_repository.dart';
+import 'package:nishauri/src/features/hiv/data/services/art_appointments.dart';
+import 'package:nishauri/src/features/lab/data/repository/ViralLoadRepository.dart';
+import 'package:nishauri/src/features/lab/data/services/ViralLoadService.dart';
 import 'package:nishauri/src/features/user/data/respositories/UserRepository.dart';
 import 'package:nishauri/src/shared/models/token_pair.dart';
 import 'dart:developer' as developer;
@@ -9,6 +17,10 @@ class AuthController extends StateNotifier<AsyncValue<AuthState>> {
   /// Act as the view model that hold the state of component or screen
   final AuthRepository _repository;
   final UserRepository _userRepository;
+
+  ProgramRepository programRepository = ProgramRepository(ProgramService());
+  ViralLoadRepository viralLoadRepository = ViralLoadRepository(ViralLoadService());
+  AppointmentRepository appointmentRepository = AppointmentRepository(AppointmentService(), false, ARTAppointmentRepository(ARTAppointmentService()));
 
   AuthController(this._repository, this._userRepository) : super(const AsyncValue.loading()) {
     loadAuthState();
@@ -41,7 +53,12 @@ class AuthController extends StateNotifier<AsyncValue<AuthState>> {
         refreshToken: authResponse.refreshToken ?? '',
       ));
       await _repository.saveUserId(authResponse.userId??'');
+      await _repository.savePhoneNumber(authResponse.phoneNumber ?? '');
       await _repository.saveIsVerified(authResponse.accountVerified);
+      await appointmentRepository.saveAppointment();
+      await _userRepository.saveGenderAge();
+      await programRepository.saveRegimen();
+      await viralLoadRepository.saveViralLoad();
       state = AsyncValue.data(
         AuthState(
           isAccountVerified: authResponse.accountVerified,
@@ -50,6 +67,26 @@ class AuthController extends StateNotifier<AsyncValue<AuthState>> {
 
         ),
       );
+  }
+  Future<bool> unlock(Map<String, dynamic> credentials) async {
+    var username = {"user_name" : await _repository.getUserPhoneNumber()};
+    final unlockCredentials = {...credentials, ...username};
+      final authResponse = await _repository.authenticate(unlockCredentials);
+      var msg = authResponse.accessToken ?? '';
+      if (msg.isEmpty){
+        throw authResponse.message??'';
+      }
+      await _repository.saveToken(TokenPair(
+        accessToken: authResponse.accessToken ?? '',
+        refreshToken: authResponse.refreshToken ?? '',
+      ));
+      await _repository.saveUserId(authResponse.userId??'');
+      await _repository.savePhoneNumber(authResponse.phoneNumber ?? '');
+      await _repository.saveIsVerified(authResponse.accountVerified);
+      await appointmentRepository.saveAppointment();
+      await _userRepository.saveGenderAge();
+      await viralLoadRepository.saveViralLoad();
+      return msg.isNotEmpty;
   }
 
   Future<void> register(Map<String, dynamic> data) async {
@@ -64,7 +101,11 @@ class AuthController extends StateNotifier<AsyncValue<AuthState>> {
         refreshToken: authResponse.refreshToken?? '',
       ));
       await _repository.saveUserId(authResponse.userId??'');
+      await _repository.savePhoneNumber(authResponse.phoneNumber ?? '');
       await _repository.saveIsVerified(authResponse.accountVerified);
+      await appointmentRepository.saveAppointment();
+      await _userRepository.saveGenderAge();
+      await viralLoadRepository.saveViralLoad();
       state = AsyncValue.data(AuthState(
         isAccountVerified: authResponse.accountVerified,
         isProfileComplete: authResponse.profileUpdated,
@@ -107,8 +148,17 @@ class AuthController extends StateNotifier<AsyncValue<AuthState>> {
   }
 
   Future<void> logout() async {
-    _repository.deleteToken();
-    _repository.deleteUserId();
+    var resp = await _userRepository.revokeToken();
+    print(resp);
+    if (resp == "Logout Successful") {
+      _repository.deleteToken();
+      _repository.deleteUserId();
+      _repository.deletePhoneNumber();
+    }
+    _userRepository.revokeToken();
+    print(_userRepository.revokeToken());
+    // _repository.deleteToken();
+    // _repository.deleteUserId();
     state.when(
       data: (value) => state = AsyncValue.data(
         value.copyWith(
